+++ src/main/java/de/marhali/easyi18n/settings/ProjectSettingsState.java
@@ -1,12 +1,12 @@
 package de.marhali.easyi18n.settings;
 
-import com.google.common.base.CaseFormat;
 import com.intellij.util.xmlb.annotations.Property;
 
 import de.marhali.easyi18n.io.parser.ParserStrategyType;
 import de.marhali.easyi18n.io.folder.FolderStrategyType;
 import de.marhali.easyi18n.settings.presets.DefaultPreset;
 
+import de.marhali.easyi18n.settings.presets.NamingConvention;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
