+++ src/main/java/de/marhali/easyi18n/settings/presets/ReactI18NextPreset.java
@@ -1,9 +1,7 @@
 package de.marhali.easyi18n.settings.presets;
 
-import com.google.common.base.CaseFormat;
 import de.marhali.easyi18n.io.parser.ParserStrategyType;
 import de.marhali.easyi18n.io.folder.FolderStrategyType;
-import de.marhali.easyi18n.settings.NamingConvention;
 import de.marhali.easyi18n.settings.ProjectSettings;
 
 import org.jetbrains.annotations.NotNull;
