+++ src/main/java/de/marhali/easyi18n/settings/presets/VueI18nPreset.java
@@ -2,7 +2,6 @@ package de.marhali.easyi18n.settings.presets;
 
 import de.marhali.easyi18n.io.parser.ParserStrategyType;
 import de.marhali.easyi18n.io.folder.FolderStrategyType;
-import de.marhali.easyi18n.settings.NamingConvention;
 import de.marhali.easyi18n.settings.ProjectSettings;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
