+++ src/main/resources/messages.properties
@@ -64,7 +64,6 @@ settings.experimental.flavor-template =I18n flavor template
 settings.experimental.flavor-template-tooltip = Specify how to replace strings with i18n representation.
 settings.experimental.key-naming-format.title=Key format of extracted translation
 settings.experimental.key-naming-format.tooltip=Choose Naming Convention for the keys of extracted translation
-settings.experimental.key-naming-format.items=Camel Case;Snake Case
 
 error.io=An error occurred while processing translation files. \n\
   Config: {0} => {1} ({2}) \n\
