+++ src/test/java/de/marhali/easyi18n/settings/SettingsTestPreset.java
@@ -1,9 +1,9 @@
 package de.marhali.easyi18n.settings;
 
-import com.google.common.base.CaseFormat;
 import de.marhali.easyi18n.io.folder.FolderStrategyType;
 import de.marhali.easyi18n.io.parser.ParserStrategyType;
 
+import de.marhali.easyi18n.settings.presets.NamingConvention;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
